CHANGE: calorieCount data field in SQLite database changed from Float to Int32. Code logic updated to support this.

+++ iCalorieTracker/View/AddCalorieItemView.swift
@@ -51,7 +51,7 @@ struct AddCalorieItemView: View {
 
                 HStack {
                     Button(action: {
-                        calorieItemListViewModel.saveCalorieItem(title: calorieTitle, id: UUID(), calorieCount: Float(calorieCount) ?? 0.0, viewContext: viewContext)
+                        calorieItemListViewModel.saveCalorieItem(title: calorieTitle, id: UUID(), calorieCount: Int32(calorieCount) ?? 0, viewContext: viewContext)
                         self.isPresented = false
                     })
                     {
@@ -73,7 +73,7 @@ struct AddCalorieItemView: View {
 
                     Spacer(minLength: 10)
                     Button(action: {
-                        calorieItemListViewModel.saveCalorieItem(title: calorieTitle, id: UUID(), calorieCount: Float(calorieCount) ?? 0.0, viewContext: viewContext)
+                        calorieItemListViewModel.saveCalorieItem(title: calorieTitle, id: UUID(), calorieCount: Int32(calorieCount) ?? 0, viewContext: viewContext)
                         self.isPresented = false
                     }) {
                         VStack{
